requirements.txt
dotenv==0.9.9
et_xmlfile==2.0.0
openpyxl==3.1.5
pdf2image==1.17.0
pillow==11.1.0
python-dotenv==1.0.1
